requirements.txt
Flask
python-docx
translators
requests
beautifulsoup4
PyMuPDF
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
google-generativeai
